+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Changes to Calva.
 
 ## [Unreleased]
 
+- Fix: [autoConnectRepl tried to connect even if no .nrepl-port present](https://github.com/BetterThanTomorrow/calva/issues/3278)
+
 ## [2.0.597] - 2026-08-27
 
 - Fix: [api: Shadow client list changes do not notify `repl.onSessionsChanged`](https://github.com/BetterThanTomorrow/calva/issues/3279)